print fibonacci sequence

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,7 +11,7 @@ void print_fibonacci(int n);
 int main(void)
 {
 	print_fibonacci(100);
-	return(0);
+	return (0);
 }
 
 /**
@@ -22,9 +22,9 @@ int main(void)
 
 void print_fibonacci(int n)
 {
-	static int first_no = 0;
-	static int second_no = 1;
-	static int next_fib;
+	static long long first_no = 0;
+	static long long second_no = 1;
+	static long long next_fib;
 
 	if (n > 0)
 	{
@@ -32,7 +32,7 @@ void print_fibonacci(int n)
 		first_no = second_no; /* for the next pass */
 		second_no = next_fib; /* for the next pass */
 
-		printf("%d, ", next_fib);
+		printf("%lld, ", next_fib);
 
 		print_fibonacci(n - 1); /* Go to next term */
 	}
